requirements.txt
appdirs==1.4.4
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
docopt==0.6.2
executing==2.0.1
# fastjsonschema==2.19.1
filelock==3.13.1
fsspec==2023.12.2
ftfy==6.1.3
gitdb==4.0.11
GitPython==3.1.41
# huggingface-hub==0.20.3
idna==3.6
importlib_metadata==7.1.0
# ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==2.1.4
matplotlib-inline==0.1.6
mistune==3.0.2
mpmath==1.3.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
# nvidia-cublas-cu12==12.1.3.1
# nvidia-cuda-cupti-cu12==12.1.105
# nvidia-cuda-nvrtc-cu12==12.1.105
# nvidia-cuda-runtime-cu12==12.1.105
# nvidia-cudnn-cu12==8.9.2.26
# nvidia-cufft-cu12==11.0.2.54
# nvidia-curand-cu12==10.3.2.106
# nvidia-cusolver-cu12==11.4.5.107
# nvidia-cusparse-cu12==12.1.0.106
# nvidia-nccl-cu12==2.19.3
# nvidia-nvjitlink-cu12==12.3.101
# nvidia-nvtx-cu12==12.1.105
packaging==23.2
pandas==2.2.0
# pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
# pipreqs==0.5.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.8
ptyprocess==0.7.0
# pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.8.2
pytz==2023.4
PyYAML==6.0.1
pyzmq==26.0.0
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rpds-py==0.18.0
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.12.0
sentence-transformers==2.3.1
sentencepiece==0.1.99
sentry-sdk==1.40.2
setproctitle==1.3.3
six==1.16.0
# smmap==5.0.1
soupsieve==2.5
# stack-data==0.6.3
sympy==1.12
threadpoolctl==3.2.0
tinycss2==1.2.1
tokenizers==0.15.1
# torch==2.2.0
# torchaudio==2.2.0
# torchvision==0.17.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.2
transformers==4.37.2
# triton==2.2.0
typing_extensions==4.9.0
# tzdata==2023.4
urllib3==2.2.0
wandb==0.16.3
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
# zipp==3.18.1